requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
requests==2.32.3
cryptography==44.0.1
python-dotenv==1.0.1
jinja2>=3.1.4
python-dateutil>=2.9.0

# Google Calendar
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1

# LLM (OpenAI-compatible — works with Groq, Ollama, etc.)
openai>=1.30.0
